requirements.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.2
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
attrs==24.2.0
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
distlib==0.3.8
Django==5.1
django-filter==1.1.0
djangorestframework==3.15.2
djangorestframework-filters==0.11.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
drf-yasg==1.21.7
filelock==3.15.4
frozenlist==1.4.1
google-api-core==2.19.1
google-api-python-client==2.140.0
google-auth==2.33.0
google-auth-httplib2==0.2.0
google-oauth2-tool==0.0.3
googleapis-common-protos==1.63.2
httplib2==0.22.0
idna==3.7
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
multidict==6.0.5
oauth2client==4.1.3
packaging==24.1
pillow==10.4.0
platformdirs==4.2.2
proto-plus==1.24.0
protobuf==5.27.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyJWT==2.9.0
pyparsing==3.1.2
pytz==2024.1
PyYAML==6.0.2
razorpay==1.4.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
rsa==4.9
setuptools==72.2.0
six==1.16.0
sqlparse==0.5.1
stripe==10.8.0
twilio==9.2.3
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
virtualenv==20.26.3
yarl==1.9.4
